create message component compact messages
